# perf(extract): splits compiler availability and execution