… test/extract/transpile/index.spec.mjs
… test/extract/transpile/meta.spec.mjs